Repackage AI metadata under org.springframework.ai.metadata.
See #98
