+++ spring-ai-openai/src/main/java/org/springframework/ai/openai/client/OpenAiClient.java
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.ai.client.AiClient;
 import org.springframework.ai.client.AiResponse;
 import org.springframework.ai.client.Generation;
-import org.springframework.ai.openai.client.metadata.OpenAiGenerationMetadata;
+import org.springframework.ai.openai.metadata.OpenAiGenerationMetadata;
 import org.springframework.ai.prompt.Prompt;
 import org.springframework.ai.prompt.messages.Message;
 import org.springframework.ai.prompt.messages.MessageType;
